pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "realtime-gym"
version = "0.1.0"
authors = [
  { name="wenyl", email="wenyl22@mails.tsinghua.edu.cn" },
]
description = "A real-time gym for evaluating language agents in dynamic environments with time constraints."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "openai>=1.0.0",
    "transformers>=4.30.0",
    "pandas>=1.3.0",
    "numpy>=1.21.0",
    "pygame>=2.0.0",
    "Pillow>=9.0.0",
    "scipy>=1.7.0",
    "gym>=0.26.0",
    "tqdm>=4.65.0",
    "ipython>=7.0.0",
    "ipywidgets>=7.6.0",
    "imageio>=2.9.0",
    "python-dotenv>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "pre-commit>=3.0.0",
    "ty",
]

[project.scripts]
agile_eval = "realtimegym.agile_eval:main"

[project.urls]
"Homepage" = "https://github.com/wenyl22/RealtimeGym"
"Bug Tracker" = "https://github.com/wenyl22/RealtimeGym/issues"
"Documentation" = "https://bleaves.github.io/real-time-reasoning/"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
realtimegym = ["prompts/*.yaml"]

[tool.ty.src]
exclude = [
    # Third-party vendored code from Overcooked-AI (https://github.com/HumanCompatibleAI/overcooked_ai)
    # See src/realtimegym/environments/overcooked_new/THIRD_PARTY_NOTICE.md for details
    "src/realtimegym/environments/overcooked_new/**",
]

# Note: ty does not currently have a rule to disallow implicit Any types
# All type annotations should be added manually for better type safety

[tool.ruff]
select = ["ANN"]
exclude = [
    # Third-party vendored code from Overcooked-AI (https://github.com/HumanCompatibleAI/overcooked_ai)
    # See src/realtimegym/environments/overcooked_new/THIRD_PARTY_NOTICE.md for details
    "src/realtimegym/environments/overcooked_new",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=src/realtimegym",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[dependency-groups]
dev = [
    "pre-commit>=3.5.0",
    "pytest>=8.3.5",
]
